widgets: split widgets in single files for better browsability
