+++ circleguard/widgets/push_button.py
@@ -1,0 +1,10 @@
+from PyQt6.QtWidgets import QPushButton
+from PyQt6.QtGui import QCursor
+from PyQt6.QtCore import Qt
+
+
+# we want most of our clickable widgets to have a pointing hand cursor on hover
+class PushButton(QPushButton):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.setCursor(QCursor(Qt.CursorShape.PointingHandCursor))
